+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [2.0.0]
+### Breaking Changes
+- Updated decrypt processor to use [crypto.createCipheriv](https://nodejs.org/api/crypto.html#crypto_crypto_createcipheriv_algorithm_key_iv_options). Rather than suppying an algorithm and password, now an algorithm, key and initialization vector iare required.
 ### Updated
 - Update dependencies
 - Remove lodash.noop dependency